Split tests into 4
