+++ packages/date-publication/__tests__/android/date-publication-bst-non-london.android.test.js
@@ -1,7 +1,5 @@
 import shared from "../shared-non-london";
 
 const dateBST = "2017-01-01T14:32:00.000Z";
-const dateGMT = "2017-07-01T14:32:00.000Z";
 
 shared(dateBST);
-shared(dateGMT);
