+++ packages/date-publication/__tests__/android/date-publication-bst.android.test.js
@@ -1,0 +1,5 @@
+import shared from "../shared-london";
+
+const dateBST = "2017-01-01T14:32:00.000Z";
+
+shared(dateBST);
