+++ packages/date-publication/__tests__/android/date-publication-gmt.android.test.js
@@ -1,7 +1,5 @@
 import shared from "../shared-london";
 
-const dateBST = "2017-01-01T14:32:00.000Z";
 const dateGMT = "2017-07-01T14:32:00.000Z";
 
-shared(dateBST);
 shared(dateGMT);
