+++ packages/date-publication/__tests__/ios/date-publication-bst-non-london.ios.test.js
@@ -1,0 +1,6 @@
+import shared from "../shared-non-london";
+
+const dateBST = "2017-01-01T14:32:00.000Z";
+
+shared(dateBST);
+
