+++ packages/date-publication/__tests__/ios/date-publication-gmt-non-london.ios.test.js
@@ -1,0 +1,6 @@
+import shared from "../shared-non-london";
+
+const dateGMT = "2017-07-01T14:32:00.000Z";
+
+shared(dateGMT);
+
