+++ packages/date-publication/__tests__/ios/date-publication-gmt.ios.test.js
@@ -1,0 +1,5 @@
+import shared from "../shared-london";
+
+const dateGMT = "2017-07-01T14:32:00.000Z";
+
+shared(dateGMT);
