+++ packages/date-publication/__tests__/web/date-publication-gmt.web.test.js
@@ -1,11 +1,9 @@
-import shared from "../shared-non-london";
+import shared from "../shared-london";
 
 const realIntl = Intl;
 
-const dateBST = "2017-01-01T14:32:00.000Z";
 const dateGMT = "2017-07-01T14:32:00.000Z";
 
-shared(dateBST);
 shared(dateGMT);
 
 global.Intl = realIntl;
